refactor(website): move blog + changelog logic into modules/, follow webjs's own convention
The website was stuffing file-reading, frontmatter parsing, and markdown
rendering directly inside `app/blog/page.ts`, `app/blog/[slug]/page.ts`,
and `app/changelog/page.ts`. That violates the layout we tell every
scaffolded webjs app to follow in AGENTS.md:

  app/                      ROUTING ONLY. Thin route adapters.
  modules/<feature>/        Feature-scoped queries + utils + types.
  lib/                      App-wide helpers.

Dogfooding fix. New layout:

  website/lib/frontmatter.ts                    Shared parser (browser-safe).
  website/modules/blog/queries/
    list-posts.server.ts                        Read all posts, return metadata.
    get-post.server.ts                          Read one post by slug, return body.
  website/modules/blog/utils/render-post.ts     Long-form markdown renderer.
  website/modules/blog/types.ts                 Post / PostWithBody.
  website/modules/changelog/queries/
    list-entries.server.ts                      Read all changelog entries.
  website/modules/changelog/utils/
    render-entry.ts                             Compact-card markdown renderer.
    pkg-badge.ts                                Color-coded package pill.
  website/modules/changelog/types.ts            Entry.

The route files at `website/app/blog/page.ts`,
`website/app/blog/[slug]/page.ts`, and `website/app/changelog/page.ts`
are now thin adapters that import from the modules and render the
result. None of them do file IO or string-parsing directly.

Same routes, same output. Logic moved to where AGENTS.md says it
should live.

The `'use server'` directive on each query file makes the file
source-protected (browser imports get a throw-at-load stub) and
RPC-callable (so a client component could in principle import
`listPosts` if it needed to, and the dev server would rewrite the
import into an RPC stub). For the current pages, both query files
are only called server-side from the page's default export.

Sharing across the two features: `parseFrontmatter()` is identical
for both, so it lives in `website/lib/frontmatter.ts` (lib/ scope
because it's cross-feature). The inline-markdown regexes diverge
between the two renderers (different code-block sizes, different
heading typography), so each module has its own renderer rather
than parameterizing a shared one. Two callers, two short
implementations, no premature abstraction.
