+++ website/app/blog/[slug]/page.ts
@@ -1,154 +1,20 @@
 import { html, unsafeHTML, notFound } from '@webjsdev/core';
-import { readFile } from 'node:fs/promises';
-import { join, dirname, resolve } from 'node:path';
-import { fileURLToPath } from 'node:url';
+import { getPost } from '../../../modules/blog/queries/get-post.server.ts';
+import { renderPostBody } from '../../../modules/blog/utils/render-post.ts';
 
 /**
  * /blog/[slug]
  *
- * Reads `blog/<slug>.md` at SSR time, parses frontmatter, renders body.
- * Returns 404 via `notFound()` if the slug does not exist.
+ * Thin route adapter. File-reading, frontmatter parsing, and markdown
+ * rendering live in `modules/blog/`. This page composes them.
  *
- * `generateMetadata` derives <head> from the post's frontmatter so each
- * post gets its own title / description / og:* tags for SEO. Each post
- * has a self-contained canonical URL at `/blog/<slug>`, matching what
- * search engines index.
- *
- * Spacing utilities use arbitrary values (e.g. `mt-[5rem]`) instead of
- * Tailwind's named scale (`mt-20`) because the Tailwind watcher in the
- * dev server is not always running while the blog/ directory is being
- * edited, so the named-scale classes can be missing from the compiled
- * output. Arbitrary-value classes are emitted from the literal token in
- * source, so they get compiled regardless.
+ * `generateMetadata` derives <head> from the post's frontmatter so
+ * each post gets its own title / description / og:* tags for SEO.
+ * Canonical URL per post at `/blog/<slug>`.
  */
 
-const __dirname = dirname(fileURLToPath(import.meta.url));
-const REPO_ROOT = resolve(__dirname, '..', '..', '..', '..');
-const BLOG_DIR = resolve(REPO_ROOT, 'blog');
-
-function parseFrontmatter(raw: string): { fm: Record<string, string>; body: string } {
-  const m = raw.match(/^---\n([\s\S]*?)\n---\n?([\s\S]*)$/);
-  if (!m) return { fm: {}, body: raw };
-  const fm: Record<string, string> = {};
-  for (const line of m[1].split('\n')) {
-    const idx = line.indexOf(':');
-    if (idx < 0) continue;
-    const k = line.slice(0, idx).trim();
-    let v = line.slice(idx + 1).trim();
-    if (v.startsWith('"') && v.endsWith('"')) v = v.slice(1, -1);
-    fm[k] = v;
-  }
-  return { fm, body: m[2] };
-}
-
-async function loadPost(slug: string) {
-  if (!/^[a-z0-9-]+$/.test(slug)) return null;
-  let raw: string;
-  try { raw = await readFile(join(BLOG_DIR, slug + '.md'), 'utf8'); }
-  catch { return null; }
-  const { fm, body } = parseFrontmatter(raw);
-  if (!fm.title || !fm.date) return null;
-  return {
-    slug,
-    title: fm.title,
-    date: fm.date,
-    description: fm.description || '',
-    tags: (fm.tags || '').split(',').map((t) => t.trim()).filter(Boolean),
-    author: fm.author || 'Vivek',
-    body: body.trim(),
-  };
-}
-
-function inline(s: string): string {
-  let out = s
-    .replace(/&/g, '&amp;')
-    .replace(/</g, '&lt;')
-    .replace(/>/g, '&gt;');
-  out = out.replace(/\[([^\]]+)\]\(([^)]+)\)/g, (_m, t, u) =>
-    `<a href="${u}" class="text-accent no-underline hover:underline" rel="noopener noreferrer">${t}</a>`);
-  out = out.replace(/`([^`]+)`/g, '<code class="font-mono text-[0.9em] bg-bg-subtle text-fg px-[6px] py-[2px] rounded">$1</code>');
-  out = out.replace(/\*\*([^\n]+?)\*\*/g, '<strong class="font-semibold text-fg">$1</strong>');
-  out = out.replace(/(^|[^\w])_([^_\s][^_]*[^_\s]|[^_\s])_(?=$|[^\w])/g, '$1<em>$2</em>');
-  out = out.replace(/(^|[^*\w])\*([^*\s][^*]*[^*\s]|[^*\s])\*(?=$|[^*\w])/g, '$1<em>$2</em>');
-  return out;
-}
-
-/** Render body. */
-function renderBody(md: string): string {
-  const lines = md.split('\n');
-  const out: string[] = [];
-  let inList = false;
-  let curItem: string[] = [];
-  let inCode = false;
-  let codeBuf: string[] = [];
-  let codeLang = '';
-
-  function flushItem() {
-    if (curItem.length) {
-      out.push(`<li class="text-fg-muted text-[17px] leading-[1.8] relative pl-[28px] my-[12px] before:content-['•'] before:absolute before:left-[6px] before:top-0 before:text-fg-subtle before:font-bold">${inline(curItem.join(' '))}</li>`);
-      curItem = [];
-    }
-  }
-  function endList() {
-    flushItem();
-    if (inList) { out.push('</ul>'); inList = false; }
-  }
-  function startList() {
-    if (!inList) { out.push('<ul class="my-[32px] space-y-[8px] list-none ml-[8px]">'); inList = true; }
-  }
-
-  for (const raw of lines) {
-    if (inCode) {
-      if (raw.trim().startsWith('```')) {
-        const escaped = codeBuf.join('\n')
-          .replace(/&/g, '&amp;').replace(/</g, '&lt;').replace(/>/g, '&gt;');
-        out.push(`<pre class="bg-bg-subtle border border-border rounded-lg my-[48px] overflow-x-auto"><code class="font-mono text-[13px] leading-[1.7] text-fg whitespace-pre block px-[24px] py-[20px]"${codeLang ? ` data-lang="${codeLang}"` : ''}>${escaped}</code></pre>`);
-        codeBuf = [];
-        codeLang = '';
-        inCode = false;
-      } else {
-        codeBuf.push(raw);
-      }
-      continue;
-    }
-    if (raw.trim().startsWith('```')) {
-      endList();
-      inCode = true;
-      codeLang = raw.trim().slice(3).trim();
-      continue;
-    }
-    const line = raw;
-    if (/^# /.test(line)) {
-      endList();
-      out.push(`<h2 class="font-serif text-[clamp(26px,3.5vw,34px)] leading-[1.18] tracking-tight text-fg mt-[80px] mb-[24px]">${inline(line.slice(2).trim())}</h2>`);
-    } else if (/^## /.test(line)) {
-      endList();
-      out.push(`<h3 class="font-serif text-[clamp(21px,2.8vw,26px)] leading-[1.2] tracking-tight text-fg mt-[56px] mb-[20px]">${inline(line.slice(3).trim())}</h3>`);
-    } else if (/^### /.test(line)) {
-      endList();
-      out.push(`<h4 class="font-mono text-[12px] uppercase tracking-[0.18em] font-semibold text-fg-subtle mt-[40px] mb-[16px]">${inline(line.slice(4).trim())}</h4>`);
-    } else if (/^> /.test(line)) {
-      endList();
-      out.push(`<blockquote class="border-l-2 border-accent pl-[20px] my-[40px] italic text-fg-muted text-[17px] leading-[1.7]">${inline(line.slice(2).trim())}</blockquote>`);
-    } else if (/^- /.test(line)) {
-      flushItem();
-      startList();
-      curItem.push(line.slice(2).trim());
-    } else if (inList && /^ {2,}\S/.test(line)) {
-      curItem.push(line.trim());
-    } else if (line.trim() === '') {
-      flushItem();
-    } else {
-      endList();
-      out.push(`<p class="text-fg-muted text-[17px] leading-[1.8] my-[28px]">${inline(line.trim())}</p>`);
-    }
-  }
-  endList();
-  return out.join('\n');
-}
-
 export async function generateMetadata({ params }: { params: { slug: string } }) {
-  const post = await loadPost(params.slug);
+  const post = await getPost(params.slug);
   if (!post) return { title: 'Post not found · webjs' };
   return {
     title: `${post.title} · webjs blog`,
@@ -167,7 +33,7 @@ export async function generateMetadata({ params }: { params: { slug: string } })
 }
 
 export default async function BlogPost({ params }: { params: { slug: string } }) {
-  const post = await loadPost(params.slug);
+  const post = await getPost(params.slug);
   if (!post) notFound();
 
   return html`
@@ -188,7 +54,7 @@ export default async function BlogPost({ params }: { params: { slug: string } })
         <p class="text-fg-muted text-[19px] leading-[1.55] m-0 font-serif italic">${post.description}</p>
       </header>
 
-      <article class="mt-[16px]">${unsafeHTML(renderBody(post.body))}</article>
+      <article class="mt-[16px]">${unsafeHTML(renderPostBody(post.body))}</article>
 
       <footer class="mt-[128px] pt-[40px] border-t border-border">
         <a href="/blog" class="font-mono text-[12px] text-fg-subtle no-underline hover:text-fg tracking-wide">← All posts</a>
