+++ website/modules/blog/types.ts
@@ -1,0 +1,12 @@
+export type Post = {
+  slug: string;
+  title: string;
+  date: string;
+  description: string;
+  tags: string[];
+  author: string;
+};
+
+export type PostWithBody = Post & {
+  body: string;
+};
